+++ gpaw/test/xc/qna_spinpol.py
@@ -9,7 +9,8 @@ QNA = {'alpha': 2.0,
        'setup_name': 'PBE',
        'type': 'qna-gga'}
 
-atoms = BodyCenteredCubic(symbol='Fe', latticeconstant=2.854,
+atoms = BodyCenteredCubic(symbol='Fe',
+                          latticeconstant=2.854,
                           pbc=(1, 1, 1))
 
 atoms.set_initial_magnetic_moments([2, 2])
@@ -24,6 +25,6 @@ atoms.set_calculator(calc)
 atoms.get_potential_energy()
 magmoms = atoms.get_magnetic_moments()
 
-tol = 1e-5
-equal(2.253246, magmoms[0], tol)
-equal(2.253246, magmoms[1], tol)
+tol = 0.003
+equal(2.243, magmoms[0], tol)
+equal(2.243, magmoms[1], tol)