+++ gpaw/test/xc/tb09.py
@@ -1,12 +1,11 @@
 """Test Tran Blaha potential."""
 from ase.dft.bandgap import bandgap
 from ase.build import bulk
-from gpaw import GPAW, PW, Davidson, Mixer
+from gpaw import GPAW, PW, Mixer
 
 k = 8
 atoms = bulk('Si')
 atoms.calc = GPAW(mode=PW(300),
-                  eigensolver=Davidson(2),
                   mixer=Mixer(0.8, 10, 50.0),
                   kpts={'size': (k, k, k), 'gamma': True},
                   xc='TB09',